[perspectiveTaking] Fixing critical bug, some other small changes Critical bug where point cloud was randomly not shown fixed. This was due to setting the camera position in a different thread from the spin, whereas now the camera position is set just before the spinning Using try_lock instead of lock for responsive drawing Added a mutex for CameraKinectWrapper, just to be safe Deleted non-used variables Remove ABM Image Provider when closing the module Fix in setting the background color (double instead of rgb is expected) Using dynamic casts instead of explicit casts Simplified updateCameraposition

+++ main/src/modules/perspectiveTaking/include/VisualizerWrapper.h
@@ -48,7 +48,7 @@ public:
 
     const std::map<std::string, boost::shared_ptr<cloudWithPose> > & getAddedClouds() const {return _addedClouds;}
 
-    void setBackgroundColor(const int r, const int g, const int b) {
+    void setBackgroundColor(const double r, const double g, const double b) {
         _visualizer->setBackgroundColor(r, g, b);
     }
